Replace lpil/Louis Pilfold with contributors
No copyright assignment agreement exists in this repo; thus each individual contribution holds the copyright over each piece of their own code, so the copyright statement should reflect this.
Also, the license should not be edit but be presented as is, verbatim. The notice example is a template for how to show it in a program, not one to be edited in the license.